Se agrega manejo de errores HTTPS

+++ src/users/users.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, HttpException, HttpStatus } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { User } from './user.entity';
 import { Repository } from 'typeorm';
@@ -11,7 +11,20 @@ export class UsersService {
     @InjectRepository(User) private userRepository: Repository<User>,
   ) {}
 
-  createUser(user: CreateUserDto) {
+  async createUser(user: CreateUserDto) {
+    const userFound = await this.userRepository.findOne({
+      where: {
+        username: user.username,
+      },
+    });
+
+    if (userFound) {
+      return new HttpException(
+        'Nombre de usuario ya existe',
+        HttpStatus.CONFLICT,
+      );
+    }
+
     const newUser = this.userRepository.create(user);
     return this.userRepository.save(newUser);
   }
@@ -20,19 +33,52 @@ export class UsersService {
     return this.userRepository.find();
   }
 
-  getUser(id: number) {
-    return this.userRepository.findOne({
+  async getUser(id: number) {
+    const userFound = await this.userRepository.findOne({
       where: {
-        id: id,
+        id,
       },
     });
+
+    if (!userFound) {
+      return new HttpException('Usuario no existe', HttpStatus.NOT_FOUND);
+    }
+    return userFound;
   }
 
-  deleteUser(id: number) {
+  async deleteUser(id: number) {
+    const userFound = await this.userRepository.findOne({
+      where: {
+        id,
+      },
+    });
+
+    if (!userFound) {
+      return new HttpException('Usuario no encontrado', HttpStatus.NOT_FOUND);
+    }
+
     return this.userRepository.delete({ id });
+
+    //Otra forma es con:
+    //const result = await this.userRepository.delete({ id });
+    //if (result.affected === 0) {
+    // return new HttpException('Usuario no encontrado', HttpStatus.NOT_FOUND);
+    //}
+    //return result;
+    //}
   }
 
-  updateUser(id: number, user: UpdateUserDto) {
-    return this.userRepository.update({ id }, user);
+  async updateUser(id: number, user: UpdateUserDto) {
+    const userFound = await this.userRepository.findOne({
+      where: {
+        id,
+      },
+    });
+
+    if (!userFound) {
+      return new HttpException('Usuario no encontrado', HttpStatus.NOT_FOUND);
+    }
+    const updateUser = Object.assign(userFound, user);
+    return this.userRepository.save(updateUser);
   }
 }
